clarify type of rule for timerEntry made from upcoming
Now this is not a functionnal change and has no impact. It is intented to clarify what is required for timerEntry made from upcoming.
NewFromTimer() try to buid a valid MythTV rule for an entry. On failure it must return a unknown type.
For particular types of entry (eg  UPCOMING ...) the type hadn't usage and 'Not_Recording' was used.
Now the right type is returned and it could be evaluated in the future.