A setting to send (original) Airdate to Kodi instead of the recording's Start Date/Time.
Useful because I almost always want to watch programs in the order that they originally aired, regardless of the order that I recorded them in.

Airdate is usually a Date (with no time)
If two shows aired on the same date, Airdate alone will not tell us which was first.
To fix that, we fall back on the recordingTime to tell us which was recorded first.  This is our best guess as to which show should become before the other.  It's an edge case, but you know....